# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comet API",
    "version": "1.0.0",
    "description": "API documentation for Comet: The Speed of Light Comic Reader."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API"
    }
  ],
  "components": {
    "schemas": {
      "ReadingProgress": {
        "type": "object",
        "required": [
          "lastPage",
          "totalPages"
        ],
        "properties": {
          "lastPage": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 1
          },
          "zoomLevel": {
            "type": "number",
            "default": 1
          },
          "readStatus": {
            "type": "string",
            "enum": [
              "UNREAD",
              "READING",
              "COMPLETED"
            ]
          },
          "timeDelta": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "ComicUpdateSchema": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "series": {
            "type": "string",
            "nullable": true
          },
          "issue": {
            "type": "integer",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "comicVineId": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ComicVineMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "issue_number": {
            "type": "string"
          },
          "cover_date": {
            "type": "string"
          },
          "image": {
            "type": "object",
            "properties": {
              "icon_url": {
                "type": "string"
              },
              "medium_url": {
                "type": "string"
              },
              "screen_url": {
                "type": "string"
              },
              "super_url": {
                "type": "string"
              },
              "thumb_url": {
                "type": "string"
              },
              "tiny_url": {
                "type": "string"
              },
              "original_url": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/comics/{id}": {
      "get": {
        "summary": "Get a specific comic",
        "tags": [
          "Comics"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comic retrieved successfully"
          },
          "404": {
            "description": "Comic not found"
          }
        },
        "operationId": "get_comics_id"
      },
      "patch": {
        "summary": "Update comic metadata",
        "tags": [
          "Comics"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComicUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comic updated successfully"
          },
          "404": {
            "description": "Comic not found"
          }
        },
        "operationId": "patch_comics_id"
      },
      "delete": {
        "summary": "Delete comic",
        "tags": [
          "Comics"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comic deleted successfully"
          },
          "404": {
            "description": "Comic not found"
          }
        },
        "operationId": "delete_comics_id"
      }
    },
    "/comics/{id}/progress": {
      "put": {
        "summary": "Update reading progress",
        "tags": [
          "Comics"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingProgress"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Progress updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "put_comics_id_progress"
      }
    },
...
    "/comics/{id}/favorite": {
      "post": {
        "summary": "Add to favorites",
        "tags": [
          "Comics"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Added to favorites"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "post_comics_id_favorite"
      },
      "delete": {
        "summary": "Remove from favorites",
        "tags": [
          "Comics"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed from favorites"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "delete_comics_id_favorite"
      }
    },
    "/comics/search-metadata": {
      "get": {
        "summary": "Search ComicVine metadata",
        "tags": [
          "Comics"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComicVineMetadata"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "get_comics_search-metadata"
      }
    },
...
  }
}